third_party/blink/web_tests/fast/html/object-border-expected.txt

Tests for border attribute with Object tag
Bug 48596 : HTML5 Conformance Test failure: approved/xhtml5/object_border_pixel.xhtml

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Title	Actual Image	Expected Image

Object Tag with border="50" px
		
PASS border-top-width is 50px.
PASS border-left-width is 50px.
PASS border-bottom-width is 50px.
PASS border-right-width is 50px.
PASS successfullyParsed is true

TEST COMPLETE
